Move everything from FeedScreen into PostingListScreen
Also delete FeedScreen from project.
There shouldn't be any different in functionality.
